upgrade to 2.0.0-SNAPSHOT.

+++ src/main/java/fastdfs/client/FastdfsClient.java
@@ -383,7 +383,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<byte[]> download(String fileId, int offset, int size) {
+    public CompletableFuture<byte[]> download(String fileId, long offset, long size) {
         ByteArrayOutputStream out = new ByteArrayOutputStream();
         return download(fileId, out, offset, size).thenApply(res -> out.toByteArray());
     }
@@ -395,7 +395,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<Void> download(String fileId, Object out, int offset, int size) {
+    public CompletableFuture<Void> download(String fileId, Object out, long offset, long size) {
         return download(FileId.fromString(fileId), out, offset, size);
     }
 
@@ -405,7 +405,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<byte[]> download(FileId fileId, int offset, int size) {
+    public CompletableFuture<byte[]> download(FileId fileId, long offset, long size) {
         ByteArrayOutputStream out = new ByteArrayOutputStream();
         return download(fileId, out, offset, size).thenApply(res -> out.toByteArray());
     }
@@ -417,7 +417,7 @@ public final class FastdfsClient implements Closeable {
      * @param size
      * @return
      */
-    public CompletableFuture<Void> download(FileId fileId, Object out, int offset, int size) {
+    public CompletableFuture<Void> download(FileId fileId, Object out, long offset, long size) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(out, "out must not be null.");
         return trackerClient
@@ -518,7 +518,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, File file, int offset) {
+    public CompletableFuture<Void> modify(String fileId, File file, long offset) {
         return modify(FileId.fromString(fileId), file, offset);
     }
 
@@ -528,7 +528,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, File file, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, File file, long offset) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(file, "file must not be null.");
         return trackerClient
@@ -542,7 +542,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, byte[] bytes, int offset) {
+    public CompletableFuture<Void> modify(String fileId, byte[] bytes, long offset) {
         return modify(FileId.fromString(fileId), bytes, offset);
     }
 
@@ -552,7 +552,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, byte[] bytes, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, byte[] bytes, long offset) {
         return modify(fileId, bytes, bytes.length, offset);
     }
 
@@ -563,7 +563,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(String fileId, Object content, long size, int offset) {
+    public CompletableFuture<Void> modify(String fileId, Object content, long size, long offset) {
         return modify(FileId.fromString(fileId), content, size, offset);
     }
 
@@ -574,7 +574,7 @@ public final class FastdfsClient implements Closeable {
      * @param offset
      * @return
      */
-    public CompletableFuture<Void> modify(FileId fileId, Object content, long size, int offset) {
+    public CompletableFuture<Void> modify(FileId fileId, Object content, long size, long offset) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         Objects.requireNonNull(content, "content must not be null.");
         return trackerClient
@@ -605,7 +605,7 @@ public final class FastdfsClient implements Closeable {
      * @param truncatedSize
      * @return
      */
-    public CompletableFuture<Void> truncate(String fileId, int truncatedSize) {
+    public CompletableFuture<Void> truncate(String fileId, long truncatedSize) {
         return truncate(FileId.fromString(fileId), truncatedSize);
     }
 
@@ -616,7 +616,7 @@ public final class FastdfsClient implements Closeable {
      * @param truncatedSize 截取字节数
      * @return
      */
-    public CompletableFuture<Void> truncate(FileId fileId, int truncatedSize) {
+    public CompletableFuture<Void> truncate(FileId fileId, long truncatedSize) {
         Objects.requireNonNull(fileId, "fileId must not be null.");
         return trackerClient
                 .updateStorageGet(fileId)
